perf: 修改 django 版本
