+++ requirements/rpm_pkg.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 yum -y install \
     gcc-c++ sshpass mariadb-devel openldap-devel libxml2-devel \
-    xmlsec1-devel xmlsec1-openssl-devel libtool-ltdl-devel
+    xmlsec1-devel xmlsec1-openssl-devel libtool-ltdl-devel \
+    postgresql-devel
